text-detection-api/requirements.txt
fastapi
uvicorn
easyocr
numpy
opencv-python
Pillow
mlflow
python-multipart
psycopg2-binary
minio